requirements.txt
discord.py
discord.py[voice]
PyNaCl
flask
opus-api==0.6.2
